move product list query process to ProductService.js
